Median Calculation using Python
Median Calculation Using Python –

In this program, I calculated the median of a given set of numbers using Python.

First, I accepted multiple numbers from the user in a single line using the input() function. These values were converted into a list of floating-point numbers using map() and list().

Next, I sorted the list in ascending order using the sort() method. Sorting is necessary because the median depends on the position of values.

Then, I found the total number of elements using the len() function.

If the number of elements is odd, the median is the middle element of the sorted list.

If the number of elements is even, the median is calculated by taking the average of the two middle elements.

Finally, I printed the calculated median using the print() statement.

Thus, the program correctly calculates the median for both even and odd numbers of inputs.